mcp-server-pro/pyproject.toml
[project]
name = "mcp-server"
version = "1.0.0"
description = "MCP Server for IA Agent with Tools"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "pinecone>=3.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.optional-dependencies]
text = ["openai>=1.0.0"]
excel = ["xlwings>=0.30.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
mcp-server = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
